Got list of ints

+++ kotlin/Sudoku.kt
@@ -3,8 +3,8 @@ import java.io.File
 /**
  * Parses a file of integers
  */
-fun parse(file: String): List<Int> {
-    return File(file).readText().trim().split(regex="\\s+".toRegex()).map { it.toInt() }
+fun parse(file: String): List<Short> {
+    return File(file).readText().trim().split(regex="\\s+".toRegex()).map { it.toShort() }
 }
 
 fun main(args: Array<String>) {
